+++ classes/Generator/Task/Generate/Config.php
@@ -16,6 +16,7 @@ class Generator_Task_Generate_Config extends Task_Generate
 	protected $_options = array(
 		'name'    => '',
 		'values'  => '',
+		'import'  => '',
 	);
 
 	/**
@@ -49,6 +50,7 @@ class Generator_Task_Generate_Config extends Task_Generate
 		return Generator::build()
 			->add_config($options['name'])
 				->value($options['values'])
+				->import($options['import'])
 				->template($options['template'])
 				->module($options['module'])
 				->pretend($options['pretend'])