+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Generator for Kohana - Changelog
 ================================
 
+Version 1.1
+-----------
+
 - Added `--import` option for `generate:config` and `generate:message`, allows
   importing of existing values into config or message output.
 - Added get_config(), get_message() and expand_path() convenience methods to
@@ -12,7 +15,8 @@ Generator for Kohana - Changelog
   be replaced by default without needing the `--force` option.
 - Inherited methods now return the values of invoked parent methods by default
   unless the return value is known to be void.
-- Added support for absolute paths to view template files.
+- Added support for specifying absolute paths to view template files with new
+  Generator_View class.
 
 Version 1.0
 -----------